requirements.txt
langchain
langchain-community
langchain-openai
sqlalchemy
psycopg2-binary
python-dotenv